Reorganize benchmark directory
This reorganization also allows to not instantiate all benchmarks for
every type.
