+++ benchmarks/bvh_driver/CMakeLists.txt
@@ -9,7 +9,7 @@ find_package(Threads REQUIRED)
 
 add_executable(ArborX_BoundingVolumeHierarchy.exe bvh_driver.cpp)
 target_link_libraries(ArborX_BoundingVolumeHierarchy.exe ArborX::ArborX benchmark::benchmark Boost::program_options Threads::Threads)
-target_include_directories(ArborX_BoundingVolumeHierarchy.exe PRIVATE ${POINT_CLOUDS_INCLUDE_DIR} ${UNIT_TESTS_INCLUDE_DIR})
+target_include_directories(ArborX_BoundingVolumeHierarchy.exe PRIVATE ${POINT_CLOUDS_INCLUDE_DIR} ${UNIT_TESTS_INCLUDE_DIR} ${CMAKE_CURRENT_SOURCE_DIR})
 add_test(NAME ArborX_BoundingVolumeHierarchy_Benchmark COMMAND ./ArborX_BoundingVolumeHierarchy.exe --buffer=0 --benchmark_color=true)
 if(ARBORX_PERFORMANCE_TESTING)
   target_compile_definitions(ArborX_BoundingVolumeHierarchy.exe PRIVATE ARBORX_PERFORMANCE_TESTING)
